robot/requirements.txt
Faker==40.1.2
psycopg2==2.9.11
robotframework==7.4.1
robotframework-browser==19.12.3
robotframework-jsonlibrary==0.5
python-dotenv==1.2.1
